Fix partial rows in CDC to Scylla S3
Before this change, because we were writing a full-chunk at a time,
a .csv file row could be cut in middle (if it was on the boundary
of the chunk). This change still maintains writing whole chunks, but
size on the object/part is adjusted to be at the end of whole .csv
row.

Also fixed some comments and license information.